+++ src/helper/index.ts
@@ -2,6 +2,6 @@ export * from './dep-helper';
 export * from './injector-helper';
 export * from './parameter-helper';
 export * from './provider-helper';
-export * from './util';
+export * from './utils';
 export * from './is-function';
 export * from './hook-helper';
